requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
pydantic==2.6.4
python-dotenv==1.0.1

openai==1.30.1
langchain==0.1.17
langchain-core==0.1.52
langchain-community==0.0.36
faiss-cpu==1.8.0
tiktoken==0.6.0

sentence-transformers==2.7.0
httpx==0.27.0
gunicorn==21.2.0
